requirements.txt
discord.py
openai==0.19.0